<filesystem> support in device addition/removal for the Qemu driver
This commit allows to attach/detach a <filesystem> device in qemu (which would previously return an error).

For this purpose I've introduced 2 new functions "virDomainFSInsert" and "virDomainFSRemove" and
added the necessary code in the qemu driver.

It compares filesystems based on their "destination" folder. So if 2
filesystems share a same destination, they are considered equal and the
Qemu driver would reject a new insertion.